+++ calwebb_spec2_pytests/L_cube_build/test_cube_build.py
@@ -1,14 +1,14 @@
 
 """
-py.test module for unit testing the barshadow step.
+py.test module for unit testing the cube_build step.
 """
 
 import pytest
 import os
 
-from jwst.barshadow.barshadow_step import BarShadowStep
+from jwst.cube_build.cube_build_step import CubeBuildStep
 from .. import core_utils
-from . import barshadow_utils
+from . import cube_build_utils
 
 
 # Set up the fixtures needed for all of the tests, i.e. open up all of the FITS files
@@ -16,7 +16,7 @@ from . import barshadow_utils
 # Default names of pipeline input and output files
 @pytest.fixture(scope="module")
 def set_inandout_filenames(request, config):
-    step = "barshadow"
+    step = "cube_build"
     step_info = core_utils.set_inandout_filenames(step, config)
     step_input_filename, step_output_filename, in_file_suffix, out_file_suffix, True_steps_suffix_map = step_info
     return step, step_input_filename, step_output_filename, in_file_suffix, out_file_suffix, True_steps_suffix_map
@@ -27,10 +27,10 @@ def set_inandout_filenames(request, config):
 def output_hdul(set_inandout_filenames, config):
     set_inandout_filenames_info = core_utils.read_info4outputhdul(config, set_inandout_filenames)
     step, txt_name, step_input_file, step_output_file, run_calwebb_spec2, outstep_file_suffix = set_inandout_filenames_info
-    stp = BarShadowStep()
-    # Only run step if data is MOS
+    # Only run step if data is IFU
     inhdu = core_utils.read_hdrfits(step_input_file, info=False, show_hdr=False)
-    if core_utils.check_MOS_true(inhdu):
+    if core_utils.check_IFU_true(inhdu):
+        stp = CubeBuildStep()
         # if run_calwebb_spec2 is True calwebb_spec2 will be called, else individual steps will be ran
         step_completed = False
         if not run_calwebb_spec2:
@@ -50,12 +50,12 @@ def output_hdul(set_inandout_filenames, config):
                 core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
                 pytest.skip("Skipping "+step+". Step set to False in configuration file.")
     else:
-        pytest.skip("Skipping "+step+" because data is not MOS.")
+        pytest.skip("Skipping "+step+" because data is not IFU.")
 
 
 
 # Unit tests
 
-def test_s_barsha_exists(output_hdul):
-    assert barshadow_utils.s_barsha_exists(output_hdul), "The keyword S_BARSHA was not added to the header --> Barshadow step was not completed."
+def test_s_ifucub_exists(output_hdul):
+    assert cube_build_utils.s_ifucub_exists(output_hdul), "The keyword S_IFUCUB was not added to the header --> IFU cube_build step was not completed."
